page accueil encheres mylan

+++ src/main/java/projet/bid_pro/dal/UtilisateurDAO.java
@@ -4,7 +4,7 @@ import jdk.jshell.execution.Util;
 import projet.bid_pro.bo.Utilisateur;
 
 public interface UtilisateurDAO {
-	Utilisateur read(long id);
+	Utilisateur read(int id);
 
 	Utilisateur read(String email);
